refactor: cleaned up bot listeners
Split up bot listeners into separate cogs.
